Turn libsodium into a Swift module
That's basically all you need for Swift bindings. But who wants to call
C functions and what would be the point of this project else?

Some details are explained in a SO answer [1], which at the time of the
commit is unfortunately not marked as the accepted answer. Linking can
also be done with a static library (`*.a`) instead of the mentioned
`*.tbd` file.

[1] http://stackoverflow.com/a/37072619/5082444