Avoid using optionals for a programming error
The only reason the `Data(hex:)` initalizer could fail is if
`libsodium.hex2bin()` tries to read more bytes than reserved. This can
only happen if not enough memory is reserved, which is a programming
error and should not be propagated to the caller. An assertion was added
just in case.