Avoid superfluous optionals and ensure initialization of `libsodium`
In order to ensure the initialization of `libsodium` before any method
from `libsodium` is called there is now a small wrapper.

In addition there is some code that unecessarily delegated optionals to
callers although the reason for a value being optional is a programming
error. This change might result in crashes if `libsodium` functions act
differently than documented.