Use byte arrays rather than the `Data` type
This allows quite a few simplifications when calling the C-based API of
Sodium [1].

I'm not happy with every change here. There are some places where we
still use the `Data` type, e.g., in the `KeyMaterial.copyBytes()` function
or in the `GernericHash.hexValue` property.

This was inspired by a swift-sodium commit [2].

[1] https://developer.apple.com/swift/blog/?id=6
[2] https://github.com/jedisct1/swift-sodium/commit/48b08e47f0e88c5adc49cabfc3daa25df23a4937