Fix invalid ignore string passed to `hex2bin`
This was uncovered by a flaky test for String.unhexlify.

The ignore string was passed without a terminating null byte. Most of
the time this worked without problems, as the tests case that was
failing was 'DE:AD:BE:EF' while ignoring '-'. Looks like some of the
characters in the input string where not in the memory region. It only
failed if there was a colon in there. Now the terminating null byte is
added by using `String.cString(using:)`.